split - cli
