+++ src/apps/cli/list_subcommand.rs
@@ -1,0 +1,76 @@
+use crate::db;
+use crate::helpers::calculate_percentage;
+use colored::Colorize;
+use rusqlite::Connection;
+use std::collections::HashMap;
+use std::process::exit;
+
+pub fn run(conn: &Connection, category: &Option<String>, date_format: &String) {
+    let tasks = db::tasks::query_all(&conn);
+
+    let mut categories: HashMap<String, Vec<db::tasks::Task>> = HashMap::new();
+    let mut done_count: HashMap<String, usize> = HashMap::new();
+    let total_tasks: i32 = tasks.len() as i32;
+    let mut total_done = 0;
+
+    for task in tasks {
+        let key = &task.category;
+
+        if task.is_done {
+            let count = done_count.entry(key.into()).or_insert(0);
+            *count += 1;
+            total_done += 1;
+        }
+
+        categories
+            .entry(key.into())
+            .or_insert(Vec::new())
+            .push(task);
+    }
+
+    if total_tasks == 0 {
+        println!("{}", format!("No tasks found!").bright_black());
+        exit(0);
+    }
+
+    match category {
+        Some(category) => {
+            let tasks = categories.get_mut(category).unwrap_or_else(|| {
+                println!("category '{}' is not found", category);
+                exit(1);
+            });
+
+            let dones = done_count.get(category).unwrap_or(&(0 as usize));
+
+            db::tasks::print_all(category, dones, tasks, date_format);
+        }
+        None => {
+            for (key, tasks) in categories.iter_mut() {
+                let dones = done_count.get(key).unwrap_or(&(0 as usize));
+
+                db::tasks::print_all(key, dones, tasks, date_format);
+            }
+
+            println!();
+
+            println!(
+                "{}",
+                format!(
+                    "{}% of all tasks complete.",
+                    calculate_percentage(total_done, total_tasks)
+                )
+                .bright_black()
+            );
+
+            println!(
+                "{}",
+                format!(
+                    "{} done, {} undone",
+                    total_done.to_string().bright_green(),
+                    (total_tasks - total_done).to_string().bright_magenta()
+                )
+                .bright_black()
+            )
+        }
+    }
+}
